test: split test TypeScript configs
