+++ package.json
@@ -25,7 +25,11 @@
     "typecheck:infra": "tsc --noEmit -p tsconfig.infra.json",
     "typecheck:api": "tsc --noEmit -p tsconfig.api.json",
     "typecheck:webapp": "tsc --noEmit -p tsconfig.webapp.json",
-    "typecheck:test": "tsc --noEmit -p tsconfig.test.json",
+    "typecheck:test": "pnpm typecheck:test:unit && pnpm typecheck:test:integration && pnpm typecheck:test:browser && pnpm typecheck:test:e2e",
+    "typecheck:test:unit": "tsc --noEmit -p tsconfig.test.unit.json",
+    "typecheck:test:integration": "tsc --noEmit -p tsconfig.test.integration.json",
+    "typecheck:test:browser": "tsc --noEmit -p tsconfig.test.browser.json",
+    "typecheck:test:e2e": "tsc --noEmit -p tsconfig.test.e2e.json",
     "typecheck:layers": "pnpm typecheck:domain && pnpm typecheck:infra && pnpm typecheck:api && pnpm typecheck:webapp",
     "codebase:analyze": "fallow",
     "codebase:audit": "fallow audit",
