+++ test/integration/msw-server.ts
@@ -1,0 +1,1 @@
+export { server } from '../msw/server'
