+++ test/integration/setup.ts
@@ -1,0 +1,14 @@
+import { afterAll, afterEach, beforeAll } from 'vitest'
+import { server } from './msw-server'
+
+beforeAll(() => {
+  server.listen({ onUnhandledRequest: 'error' })
+})
+
+afterEach(() => {
+  server.resetHandlers()
+})
+
+afterAll(() => {
+  server.close()
+})
