+++ test/setup/integration-setup.ts
@@ -1,14 +1,1 @@
-import { afterAll, afterEach, beforeAll } from 'vitest'
-import { server } from '../msw/server'
-
-beforeAll(() => {
-  server.listen({ onUnhandledRequest: 'error' })
-})
-
-afterEach(() => {
-  server.resetHandlers()
-})
-
-afterAll(() => {
-  server.close()
-})
+import '../integration/setup'
