+++ tsconfig.json
@@ -1,41 +1,17 @@
 {
+  "extends": "./tsconfig.base.json",
   "compilerOptions": {
-    "allowJs": true,
-    "allowSyntheticDefaultImports": true,
-    "noEmit": true,
-    "declaration": false,
-    "sourceMap": true,
-    "esModuleInterop": true,
-    "inlineSourceMap": false,
     "lib": ["esnext", "DOM"],
-    "listEmittedFiles": false,
-    "listFiles": false,
-    "moduleResolution": "bundler",
-    "pretty": true,
-    "resolveJsonModule": true,
-    "rootDir": ".",
-    "skipLibCheck": true,
-    "traceResolution": false,
-    "outDir": "",
-    "target": "esnext",
-    "module": "esnext",
-    "jsx": "react-jsx",
     "types": ["vite/client"],
-    "strict": true,
-    "noImplicitAny": true,
-    "strictNullChecks": true,
-    "strictFunctionTypes": true,
-    "strictBindCallApply": true,
-    "strictPropertyInitialization": true,
-    "noImplicitThis": true,
-    "alwaysStrict": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noImplicitReturns": true,
-    "noFallthroughCasesInSwitch": true,
     "paths": {
       "@domain/*": ["./src/domain/*"],
       "@infra/*": ["./src/infra/*"],
+      "@test-utils/*": [
+        "./test/unit/*",
+        "./test/integration/*",
+        "./test/browser/*",
+        "./e2e/test-utils/*"
+      ],
       "@theme": ["./src/webapp/theme"]
     }
   },
