+++ tsconfig.test.e2e.json
@@ -1,0 +1,11 @@
+{
+  "extends": "./tsconfig.base.json",
+  "compilerOptions": {
+    "types": ["node"],
+    "paths": {
+      "@test-utils/*": ["./e2e/test-utils/*"]
+    }
+  },
+  "include": ["playwright.config.ts", "e2e/**/*.e2e.test.ts", "e2e/test-utils/**/*.ts"],
+  "exclude": ["node_modules", "dist", "build"]
+}
